Find more top level declarations
Before, if a top level declaration was referenced frome another decl,
then the loop over the Translation Unit's declarations would see that
it was already traversed, and skip it.  Thus, it was never added to the
set of top level decls.  Now, always check to make sure that a decl is
top level, even when skipping traversal.